FROST Signing Ceremony (#16)
This PR implements the onchain FROST signing ceremony coordination
logic. This allows for an established group to sign a message with a
minimum threshold of participants.

This follows very closely the RFC-9591 implementation, with a few small,
but important, differences:
- Committing to nonces is done before a signing ceremnory, by committing
to a Merkle root of a chunk of 1024 nonces. This ensures that nonces are
"set in stone" before the message to sign is revealed with is important
for preventing Wagner's Birthday Attack (whereby an attacker mines a
problematic nonce that reveals the another participants secret share).
- The coordinator implicitely assumes that the participants have some
other mechanism for computing the set `S` of signing participants. In
the context of Shieldnet, we are implicitly assuming that "all honest
participants" are part of the signing ceremony. This allows each
validator to independently compute the group commitment from the set of
revealed nonces.
- Submitting a share onchain is done with a "commitment root". This is
(another... :sob:) Merkle tree of all of the participant's commitment
shares and `c * λ` values (which each participant can compute themselves
with public information from events that are emitted when the signing
nonces are revealed). This ensures that participants can only update the
group signature values when they are using commitment shares that match
what other participants computed. This allows us to both verify
participant shares and compute the group signature onchain, without
allowing malicious participants to influence the group signature from an
honest majority.

An E2E test was added that goes over the whole signing ceremony and
successfully produces a group FROST signature that can be verified
onchain :smile:.

This superseded #15