FROST Algorithm Tests (#17)
This PR adds some test vectors from the reference Rust [FROST
crate](https://docs.rs/frost-secp256k1) to make sure the algorithms are
all correct. In particular, it found a couple issues:

1. The `nonce` generation algorithm uses a secret scalar, and not its
corresponding public key
2. The `_hash` implementation was shifting the DST offset one bit too
far, which caused the incorrect DST to be written and incorrect hashes
to be produced.
3. `keyGen` function names made consistent with the algorithm name
4. Provide and use a proper KeyGen hashing algorithm