Adjust Merkle Tree Structure for Nonces Commitments (#18)
We were enforcing a specific "sorted" Merkle tree structure for nonce
commiments (where the offset of a signature sequence would determine the
position in the Merkle tree, and our Merkle proof would used ordered
hashing, meaning the left/right-ness of the node would matter).

It turns out that this is needlessly complicated, and instead we just
include the offset in the leaf hash. This makes implementations easier
(we have a single Merkle tree structure that we use everywhere) while
still preventing the Wagner's Birthday Attack.