Epoch Rotation on Consensus Contract (#31)
This PR proposes a minimal interface for epoch rotation on the consensus
contract. It is based on our discussion to make the consensus contract
completely agnostic to _how_ epoch rotation is decided amongst the
validators.

Additionally, a view method is provided to examine the currently active
epoch and its FROST group ID.