Fix Signatures for Partially Committed Groups
Once signers have fully commited to a group (i.e. the group's public
key has been successfully computed), it can be used for signing.
Otherwise, it would be possible for a malicious participant to start a
KeyGen ceremony, start a signing ceremony with the partially resolved
group, and then "finish" the signature (i.e. make it so the contract's
`signatureVerify` function does not revert) atomically, even if other
participants were to add new group commitments later on (thereby
changing the group key and invalidating the sigature).

This PR adds a "pending" counter that tracks the number of commitments
that have happened and prevents certain operations (such as secret
sharing and starting signing ceremonies) until the group is fully
committed to.