Introduce Callback Mechanism for Coordinator
This PR introduces a callback mechanism for the FROST coordinator for
completed _KeyGen_ and _Sign_ ceremonies. This is done to facilitate
the validator implementation, in that honest validators will always
call the `*WithCallback` variants of the ceremony functions to
automatically drive the consensus contract forward.

Note that the consensus mechanism still needs to know the "last
validator" to participate in any given ceremony in order to know who
misbehaved (in case a validator did not use the `*WithCallback`
variants or called them with the wrong callback parameters). It does
reduce the number of transactions, however, needed by the validators
and generally simplifies their state machines.

In order to implement this feature, we needed to additionally track
the pending `keyGenSecretShare` calls. Instead of using an additional
storage slot, we use a marker value in the `sequence` value to
differentiate between "pending commits" and "pending secret shares".

Additionally, I needed to refactor out the `GroupId` and `SignatureId`
types into a separate file to avoid circular import issues between the
`FROSTCoordinator` contract and `IFROSTCoordinatorCallback` interface.

Driving the consensus contract with callbacks will be done in a follow
up PR.