Transaction Attestations
This PR adds support for transaction attestations to the consensus
contract. Additionally, it provides convienient methods to get the
attestations by meta-transaction and message (to pass the "Etherscan"
test of being able to easily get signatures for a transaction).

It might be a good idea to add a similar convenience method for
getting signatures for epoch rollovers, and potentially even
"building" a complete signature (that combines a meta-transaction
attestation with all the "catchup" signatures to form an epoch chain,
in other words the exact signature bytes that should be passed to a
potential guard).