Validator Specification
This PR is an attempt to formalize the validator state machine in an
unabiguous way, so that we can agree on exactly how the validator
should function.

This is PR is in draft, as I don't know whether or not we want to
commit it in its current form (especially since, it may become
outdated very quickly since we don't run any tests against it).

It is still useful to share - as we can use it as a starting point
for discussin _how_ we expect validators to work.