Validator Sqlite Storage
In order to persist validator state across restarts, we started
implementing a Sqlite based storage implementation.

For now, this JUST Implements the group storage methods with some
basic integrity unit tests.

Note that, unrelated to storage, we fix an issue with random
coefficient generation, where `randomBytes` was being used to generate
a scalar. This is unsound, as it introduces bias into the random
values which is unsafe.