State Machine Complaint Flow Specification
This PR implements the complaint flow in the state machine. Note that
we do not need to make use of the `KeyGenFinalized` event that occurs
when all participants confirmed (cc @remedcu).

Additionally, it turns out that the whole complaint flow is quite
stateless (in that you respond to the accusations right away, or you
collect the complaint responses into the existing secret share
collection logic), making it fairly easy to graft on; with very little
additional state required.