Wait for Manual KeyGen for Triggering Genesis
This PR changes the strategy for doing genesis to wait for the
`KeyGen` event to be manually triggered. With this change, we no
longer need to start indexing from block 0 and can just start with
`latest`.

Now, the genesis process looks a bit different:
1. Each validator starts with their parameters
2. Someone manually triggers a `keyGen` call on the coordinator
   contract in order to kick off the process
3. Profit $$$

This simplifies the startup significantly and no longer requires nodes
to support querying events from huge block ranges, or taking along
time to get to the current block (with the downside of a more manual
genesis process).