Parameterize Genesis Group Context
Right now, if genesis group key generation fails, the validators
require a key rotation in order to form a new group to be used for
genesis. This PR adds a new `GENESIS_SALT` configuration that can be
used to parameterize the ID of the genesis group, and allows us to
start new consensus contracts without needing to rotate validator
accounts.