Simplify Last Participant Handling (#83)
As discussed on Slack, the handling of the `last_participant` and how it
related to the `responsible` participant was unclear.

This PR simplifies handling by removing the `last_participant`
everywhere except for `share` collection. There, the last participant is
still needed (and represents the last participant to broadcast a signing
share), as it is used to determine the responsible participant for
restarting the signing ceremony in case of a timeout.