Fix Offset for Nonces Commitment (#108)
The nonces commitment chunks currently work in a way that allow
participants to commit to nonces partway through a chunk. That is, if
the signing sequence is already at `x` (meaning that `x` signing
ceremonies have already started, and the latest ceremony with a signing
sequence `x-1`), then the next `preprocess` nonces commitment is allowed
to a chunk for the current signing sequence (assuming no prior
registration). For example, let's say that the sequence is `1337`, then
it will register chunk `1` with a starting offset of `1337-1024`
assuming chunk `1` has not yet been registered. This allows the
participant to start revealing nonces and participating in signing
ceremonies right away (that is, starting from the signing sequence `x`)
without having to wait for a new chunk and recover more quickly.
However, because of Wagner's Birthday Attack, we cannot allow the
participant to join signing ceremonies with sequence `x-1` or earlier in
this case, as other signers have already revealed nonces.

While this logic is useful, there was a small bug when we were
committing for the _next_ chunk (that is, there was already a nonces
chunk commitment for the current chunk): we didn't set the offset to 0.
This means that if we were on signing sequence `1023` and had already
provided a Merkle root for chunk 0 (which sequence `1023` belongs to),
then we would add a new chunk root for chunk 1 **with starting offset of
`1023`**, making the validators not able to sign any transactions for
the first 1023 offsets of that chunk (i.e. sequences `1024-2046`). This
would make only two sequences from the new chunk valid, even if the
participant had registered that chunk's Merkle root before the chunk
started.

This PR fixes that bug, and adds a test to verify the fix.

Note that, while debugging, we had discussed that the `Commitments`
struct (`{ uint64 next; mapping(uint64 chunk => Root) chunks }`) is just
an array in disguise, and we could have re-written it as an array. I
decided to keep the current implementation because:
- So we don't need to worry about `uint64 <-> uint256` casting semantics
and related overflows.
- Avoid inefficiencies where the `length` would be read multiple times
(when incrementing while `push()`-ing, to check for out-of-bounds access
when setting, etc.).